[Upgrade] Replaced SwixTransform target invocation from BuildDependsOn property to BeforeTargets="Harvest"
The reason is that WIX 3.11.1 overwrites all properties (not appends as it is usually suggested - probably in order to replace already defined default build flow with its own), therefore BuildDependsOn approach works only in case Swix.targets is imported after Wix.

It is convenient though to import wix in the Directory.Build.targets (to avoid duplicating it in every wixproj), which is imported at a very late stage, much later than Swix's nuget package *.props and *.targets.

So BeforeTargets approach seems more robust for Swix as it avoids ordering issues completely.